General cleanup of the NPF_Ioctl function: - rewrote almost all the code issuing NdisRequest's to the underlying NDIS   driver.   In particular, added a bunch of code (NPF_StartUsingBinding()) to prevent   blue screens when the underlying Ndis handle is closing or closed, and   added code for better sync between NdisRequest and NPF_RequestComplete. - Rewrote all the tracing to use the TRACE_XXX macros. - Removed the EXIT_SUCCESS() and EXIT_FAILURE() macros in favor of   some better ones that should return more meaningful error codes. - Fixed a large number of errors with EXIT_SUCCESS(): the number of bytes   returned in Irp->IoStatus.Information was almost "random" (i.e. you return   a value different from 0 in a GET ioctl, with the actual number of bytes you   are returning). - fixed several warnings issued by PREfast. - Rewrote NPF_RequestComplete completely. Now it's much simpler.
[Merge from the AMD64 branch]
Updated the copyright to the new CACE+NetGroup one.
Fixed a DbgPrint that was enabled on release builds, too.
Cleaned up the code retrieving the capture statistics.
Added some #ifdef's to enable the generation of JITted filters in the x86-32
bit driver, only (although it's possible to port the JIT to x86-64).
Added some #ifdef's to enable the TME extensions (and the related
MONITOR mode) in the x86-32 bit driver only.
Cleaned up some comments in the code.
Fixed some casts that were causing some compilation warnings in the
x86-64 build.
Removed some duplicated local variable declarations, that were causing
PREFast warnings.
Added an initialization of a variable to NULL to avoid a PREFast warning.
Removed a useless inclusion of tme.h.