Apply the WinPcap license and a Riverbed copyright.
In each applicable WinPcap OEM source code file (.h, .c, and .cpp) apply
a BSD 3-clause license similar to the one used in the plain WinPcap
sources. This is preparation for inclusion in the public WinPcap
repository.

To the best of my knowledge Riverbed Technology has sole ownership of
each the changed files.